게시판 api 코드 boardContorller로 이동 및 pagination 스타일 적용
